switched from interface to AbstractClient abstract class for improved extensibility
